+++ NEWS.md
@@ -1,9 +1,20 @@
 
 # News
 
+
+05/20/16
+
++ `get` can now be used on attachments, returns file path
+
++ `attach` now works with attachments
+
++ `attach` supports `forgetRelations` parameter.
+
+
 05/18/16
 
 + `copy` now accepts `confirm` and `replace`
+
 + added internal `testthat` unit testing
 
 
@@ -19,6 +30,7 @@ pass to `graphics:pie` and `igraph:plot` respectively.
 + Added missing documentation for `pies`.
 
 + Fixed documentation for `attach`.
+
 
 05/03/16
 
@@ -39,6 +51,7 @@ pass to `graphics:pie` and `igraph:plot` respectively.
 field. This does not seem to be a problem. However, older items
 storing current working directory as source could be affected.
 
+
 04/28/16
 
 + now src must be an item (meant to be an attachment containing source
@@ -51,38 +64,53 @@ code). Documentation updated accordingly.
 
 + Bug fixes in lazydo and parameter check
 
+
 04/24/2016
 
 + passes `devtools::check()`
+
 + stash simplified (now `stash(x)` instead of `stash("x")`)
+
 + Fixed bug when replacing entries due to new scheme for naming files.
+
 
 01/04/2016
 
 + Bugfixes in lazydo
+
 + Minor additions to docs
 
 
 03/20/3026
 
 + Fixed warnings in cpanel
+
 + All `check()` now passed
+
 + Merged latest fix from the dev branch
+
 + The NEWS file now used for main news
 
 
 03/19/2016
 
 + Further testing
+
 + vignette updated
+
 + News moved to NEWS file
+
 + Minor release
+
 
 03/18/2016
 
 + New pull feature now working
+
 + added lazydo (run expression and cache results)
+
 + Sources auto-attach suspended (had problems)
+
 
 12/07/2015
 
@@ -92,24 +120,34 @@ code). Documentation updated accordingly.
 11/19/2015
 
 + Sources are now auto-attached
+
 + Added safe-remove of data
+
 + Added auto-attach of sources
+
 + Added notes field
+
 
 10/21/2015
 
 + Fixed regression in `check` after moving to relative paths
+
 + Added check points for reserved tags and warnings for existing tags
+
 + Improved print reaction when all matching results are hidden
+
 
 10/19/2015
 
 + Added Bulk-edit feature.
 
+
 10/17/2015
 
 + Added "Maybe-you-were-looking-for" feature to `get`.
+
 + Fixed bug with managing new relative paths.
+
 
 10/12/2015
 
@@ -118,6 +156,7 @@ rebase a repository, for example to a remote machine, without affecting
 the index.
 
 + Automatic update of old repository entries upon resource loading.
+
 
 10/08/2015